pyproject.toml
[project]
name = "langgraph-mcp-agent"
version = "0.1.0"
description = "LangGraph multi-agent system (v3 architecture) with MCP servers for tools"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "langgraph>=0.4.1",
    "langchain>=0.3.24",
    "langchain-google-genai>=2.1.0",
    "langchain-ollama>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "duckduckgo-search>=7.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
